Add files via upload
My mapreduce as well as the three files I created with it. The text files show the amount of traffic on the site subsetting on each grade-level option aggregated at the day, hour, and 10 minute levels. The R markdown file takes in the text files and cleans them, as well as reformats them into a usable form for time series.